Major changes
1) Completely changes how LogicalTopology works. Instead of multiple
PubSubGroups, there is a Client HashMap and a Broker HashMap.
a) As such, modified TopologyFileParser to reflect this change
b) Also created a new Class called ClientNotes to handle a Client's
connections and roles.
c) Made a new ClientRole enum to further separate Brokers and Clients
2) Also found an error where a node with a connection to a non-existent
node would cause an exception. Fixed that error
3) Added Distributed information
a) Created a distributed file parser
b) Modified getHost and getPort to use this distributed information.
4) Moved test files into better/proper folders