Completely changed how workloads were implemented.
Now, there is only one type of workload file that contains several
different actions instead of a publisher and subscriber type.
Furthermore, instead of client roles, each client specifies which
workload it wishes to receive in its topology file.

After all the submitted workloads are parsed, a masterWorkload is
created - containing all of the workload files and their list of
actions. When a Client is created, its workload is read off the
masterWorkload, and given to it. The Client will then run each action in
the sequence.

This meant I could remove the old Workload and ClientRole files.